Move FindTypeMeta into the proper file
